requirements.txt
streamlit>=1.20
python-dateutil
beautifulsoup4
html2text
pandas
openai
python-dotenv